Implemented light/dakk mode setting toggle
